+++ package.json
@@ -11,10 +11,9 @@
   },
   "dependencies": {
     "await-to-js": "^2.1.1",
-    "axios": "^0.19.2",
+    "axios": "^0.21.1",
     "chart.js": "^2.8.0",
     "core-js": "^3.3.2",
-    "material-design-icons": "^3.0.1",
     "moment": "^2.24.0",
     "vue": "^2.6.11",
     "vue-breakpoints": "^1.1.0",
@@ -28,7 +27,7 @@
     "vuex-persist": "^2.2.0"
   },
   "devDependencies": {
-    "@mdi/font": "^4.1.95",
+    "@mdi/font": "^5.8.55",
     "@vue/cli-plugin-babel": "^4.0.0",
     "@vue/cli-plugin-eslint": "^4.0.0",
     "@vue/cli-plugin-unit-jest": "^4.0.0",
