+++ src/components/Info/KubeConfig.vue
