+++ src/components/Info/OSInstallation.vue
