+++ src/components/Wizard/Ingress.vue
@@ -26,7 +26,7 @@
 
 <script>
 import YamlContainer from './YamlContainer'
-import EventBus from '../../plugins/eventbus.js'
+import EventBus from '@/utils/eventbus'
 
 export default {
   name: 'Ingress',
