+++ src/components/Wizard/Service.vue
@@ -20,7 +20,7 @@
 
 <script>
 import YamlContainer from './YamlContainer'
-import EventBus from '../../plugins/eventbus.js'
+import EventBus from '@/utils/eventbus'
 
 export default {
   name: 'Service',
