+++ src/main.js
@@ -1,10 +1,10 @@
 import Vue from 'vue'
 import App from './App.vue'
 import VuexPersistence from 'vuex-persist'
-import store from './store'
-import logger from '@/plugins/logger'
-import vuetify from './plugins/vuetify'
-import router from './router'
+import store from './store/store'
+import logger from './utils/logger'
+import vuetify from './utils/vuetify'
+import router from './utils/router'
 import GAuth from 'vue-google-oauth2'
 import VueClipboard from 'vue-clipboard2'
 import '@/assets/css/tailwind.css'
@@ -17,9 +17,8 @@ const gauthOption = {
 
 const vuexLocalStorage = new VuexPersistence({
   key: 'vuex', // The key to store the state on in the storage provider.
-  storage: window.localStorage
-  // reducer: (state) => ({user: state.user}) // or window.sessionStorage or localForage
-  // Function that passes the state and returns the state with only the objects you want to store.
+  storage: window.localStorage,
+  reducer: (state) => ({ users: state.users })
 }).plugin(store)
 
 Vue.use(GAuth, gauthOption)
